Conflicts:
	smtpd/smtpd.c

+++ smtpd/smtpd.c
@@ -161,6 +161,7 @@ int	profiling = 0;
 int	verbose = 0;
 int	debug = 0;
 int	foreground = 0;
+int     can_remove_socket = 0;
 
 struct tree	 children;
 
@@ -394,7 +395,10 @@ parent_shutdown(void)
 	clean_setproctitle();
 	event_base_free(NULL);
 #endif
-	unlink(SMTPD_SOCKET);
+
+	/* set by control.c:control() */
+	if (can_remove_socket)
+		unlink(SMTPD_SOCKET);
 
 	log_warnx("warn: parent terminating");
 	exit(0);
